requirements.txt
google-generativeai
streamlit
matplotlib
pandas
numpy
pytz